--- src/main/java/SimpleDriver.java
@@ -1,10 +1,0 @@
-import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.chrome.ChromeDriver;
-
-public class SimpleDriver {
-    public WebDriver getDriver() {
-        System.setProperty("webdriver.chrome.driver", "E:\\AQA18\\TAF\\src\\test\\resources\\chromedriver.exe");
-        WebDriver driver = new ChromeDriver();
-        return driver;
-    }
-}
